Take into account comdat groups when emitting cap_relocs.
If a symbol is in a comdat group and we need to emit a entry in
cap_relocs for it, emit it in a section with the same comdat
group as the original symbol. Otherwise we might end up with
references in cap_relocs to removed symbols (removed by the
linker).